GETP-211 rename: persistence.adapter.project.meeting.entity에서 persistence.adapter.project.meeting.model로 변경
