+++ src/main/java/es/princip/getp/persistence/adapter/project/meeting/ProjectMeetingPersistenceMapper.java
@@ -2,7 +2,7 @@ package es.princip.getp.persistence.adapter.project.meeting;
 
 import es.princip.getp.domain.project.meeting.model.ProjectMeeting;
 import es.princip.getp.persistence.adapter.common.mapper.PhoneNumberPersistenceMapper;
-import es.princip.getp.persistence.adapter.project.meeting.entity.ProjectMeetingJpaEntity;
+import es.princip.getp.persistence.adapter.project.meeting.model.ProjectMeetingJpaEntity;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
