+++ src/main/java/es/princip/getp/persistence/adapter/project/meeting/model/MeetingScheduleJpaVO.java
@@ -1,4 +1,4 @@
-package es.princip.getp.persistence.adapter.project.meeting.entity;
+package es.princip.getp.persistence.adapter.project.meeting.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
